requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
click==8.1.7
Deprecated==1.2.14
distro==1.8.0
frozenlist==1.4.1
fsspec==2023.12.2
greenlet==3.0.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
llama-index==0.9.15.post2
marshmallow==3.20.1
multidict==6.0.4
munkres==1.1.4
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.8.8
openai==1.4.0
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
regex==2023.10.3
SQLAlchemy==2.0.23
tenacity==8.2.3
tiktoken==0.5.2
types-protobuf==4.24.0.4
typing-inspect==0.9.0
wrapt==1.16.0
yarl==1.9.4
